Incomplete update
Automated complete audiofreq and buttonlst lists. Now I’m working on a
loop to make all the buttons. We don’t have to use every note I put up;
we can just specify a range or number of keys for the piano once I
figure out everything. Too tired to finish tonight but will work on it
as soon as I wake up